LLM-Function-Calling/requirements.txt
python-dotenv
ollama
requests